qase-cypress: fixed problem when completing tests
Cypress kills the process after the last tests. The reporter will wait for all results to be sent to Qase and will not block the  process after sending.

qase-cypress: fixed problem when completing tests

Cypress kills the process after the last tests. The reporter will wait for all results to be sent to Qase and will not block the  process after sending.